refactor(agent): replace registry service client with direct registry client usage for improved performance and clarity
